# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "TuMobilya-API",
        "description": "TuMobilya-API için kimlik doğrulama ve mobilya yönetimi API'si",
        "contact": {
            "email": "info@tumobilya.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Yerel geliştirme sunucusu"
        },
        {
            "url": "http://tumobilya-api.test"
        }
    ],
    "paths": {
...
        "/api/settings": {
            "get": {
                "tags": [
                    "Ayarlar"
                ],
                "summary": "Tüm ayarları listele",
                "description": "Kayıtlı tüm ayarları döndürür.",
                "operationId": "04b6f539c3af7d460d214177ac396d5e",
                "responses": {
                    "200": {
                        "description": "Ayarlar başarıyla listelendi",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "phone": {
                                                "type": "string",
                                                "example": "+90 555 123 45 67"
                                            },
                                            "email": {
                                                "type": "string",
                                                "example": "info@tumobilya.com"
                                            },
                                            "address": {
                                                "type": "string",
                                                "example": "Mobilya Sokak No:123, İstanbul, Türkiye"
                                            },
                                            "short_address": {
                                                "type": "string",
                                                "example": "İstanbul, Türkiye"
                                            },
                                            "facebook": {
                                                "type": "string",
                                                "example": "https://facebook.com/tumobilya"
                                            },
                                            "twitter": {
                                                "type": "string",
                                                "example": "https://twitter.com/tumobilya"
                                            },
                                            "instagram": {
                                                "type": "string",
                                                "example": "https://instagram.com/tumobilya"
                                            },
                                            "youtube": {
                                                "type": "string",
                                                "example": "https://youtube.com/tumobilya"
                                            },
                                            "general_title": {
                                                "type": "string",
                                                "example": "TuMobilya - Kaliteli Mobilyalar"
                                            },
                                            "general_description": {
                                                "type": "string",
                                                "example": "Evlerinize şıklık ve konfor katan mobilya çözümleri."
                                            },
                                            "general_photo": {
                                                "type": "string",
                                                "example": "logo.jpg"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Kimlik doğrulanmadı",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated"
                                        },
                                        "status": {
                                            "type": "integer",
                                            "example": 401
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Ayarlar"
                ],
                "summary": "Yeni ayar ekle",
                "description": "Yeni bir ayar oluşturur.",
                "operationId": "ef7c9bfa1e8e609feb26624dcf678ca1",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "phone": {
                                        "type": "string",
                                        "example": "+90 555 987 65 43"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "support@tumobilya.com"
                                    },
                                    "address": {
                                        "type": "string",
                                        "example": "Yeni Sokak No:456, İstanbul, Türkiye"
                                    },
                                    "short_address": {
                                        "type": "string",
                                        "example": "İstanbul, Türkiye"
                                    },
                                    "facebook": {
                                        "type": "string",
                                        "example": "https://facebook.com/tumobilya2"
                                    },
                                    "twitter": {
                                        "type": "string",
                                        "example": "https://twitter.com/tumobilya2"
                                    },
                                    "instagram": {
                                        "type": "string",
                                        "example": "https://instagram.com/tumobilya2"
                                    },
                                    "youtube": {
                                        "type": "string",
                                        "example": "https://youtube.com/tumobilya2"
                                    },
                                    "general_title": {
                                        "type": "string",
                                        "example": "TuMobilya - Yeni Nesil Mobilyalar"
                                    },
                                    "general_description": {
                                        "type": "string",
                                        "example": "Yeni nesil mobilya çözümleri."
                                    },
                                    "general_photo": {
                                        "type": "string",
                                        "example": "logo2.jpg"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Ayar başarıyla oluşturuldu",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "phone": {
                                                    "type": "string",
                                                    "example": "+90 555 987 65 43"
                                                },
                                                "email": {
                                                    "type": "string",
                                                    "example": "support@tumobilya.com"
                                                },
                                                "address": {
                                                    "type": "string",
                                                    "example": "Yeni Sokak No:456, İstanbul, Türkiye"
                                                },
                                                "short_address": {
                                                    "type": "string",
                                                    "example": "İstanbul, Türkiye"
                                                },
                                                "facebook": {
                                                    "type": "string",
                                                    "example": "https://facebook.com/tumobilya2"
                                                },
                                                "twitter": {
                                                    "type": "string",
                                                    "example": "https://twitter.com/tumobilya2"
                                                },
                                                "instagram": {
                                                    "type": "string",
                                                    "example": "https://instagram.com/tumobilya2"
                                                },
                                                "youtube": {
                                                    "type": "string",
                                                    "example": "https://youtube.com/tumobilya2"
                                                },
                                                "general_title": {
                                                    "type": "string",
                                                    "example": "TuMobilya - Yeni Nesil Mobilyalar"
                                                },
                                                "general_description": {
                                                    "type": "string",
                                                    "example": "Yeni nesil mobilya çözümleri."
                                                },
                                                "general_photo": {
                                                    "type": "string",
                                                    "example": "logo2.jpg"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Ayar oluşturuldu"
                                        },
                                        "status": {
                                            "type": "integer",
                                            "example": 201
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validasyon hatası",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The email field must be a valid email address."
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
...
}